+++ encoding.go
@@ -2,10 +2,70 @@ package goutils
 
 import (
 	"bytes"
+	"encoding/base64"
 	"golang.org/x/text/encoding/simplifiedchinese"
 	"golang.org/x/text/transform"
 	"io/ioutil"
+	"math/big"
 )
+
+// Base64Encoding Base64编码。
+func Base64Encoding(str string) string {
+	src := []byte(str)
+	res := base64.StdEncoding.EncodeToString(src)
+	return res
+}
+
+// Base64Decoding Base64解码。
+func Base64Decoding(str string) string {
+	res, _ := base64.StdEncoding.DecodeString(str)
+	return string(res)
+}
+
+var base58 = []byte("123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz")
+
+// Base58Encoding Base58编码。
+func Base58Encoding(str string) string {
+	strByte := []byte(str)
+
+	strTen := big.NewInt(0).SetBytes(strByte)
+
+	var modSlice []byte
+	for strTen.Cmp(big.NewInt(0)) > 0 {
+		mod := big.NewInt(0)
+		strTen58 := big.NewInt(58)
+		strTen.DivMod(strTen, strTen58, mod)
+		modSlice = append(modSlice, base58[mod.Int64()])
+	}
+	for _, elem := range strByte {
+		if elem != 0 {
+			break
+		} else if elem == 0 {
+			modSlice = append(modSlice, byte('1'))
+		}
+	}
+	ReverseModSlice := ReverseByteArr(modSlice)
+	return string(ReverseModSlice)
+}
+
+// Base58Decoding Base58解码。
+func Base58Decoding(str string) string {
+	strByte := []byte(str)
+	ret := big.NewInt(0)
+	for _, byteElem := range strByte {
+		index := bytes.IndexByte(base58, byteElem)
+		ret.Mul(ret, big.NewInt(58))
+		ret.Add(ret, big.NewInt(int64(index)))
+	}
+	return string(ret.Bytes())
+}
+
+func ReverseByteArr(bytes []byte) []byte {
+	for i := 0; i < len(bytes)/2; i++ {
+		bytes[i], bytes[len(bytes)-1-i] = bytes[len(bytes)-1-i], bytes[i]
+	}
+	return bytes
+}
 
 // GbkToUtf8 GBK vs UTF-8 编码转换。
 func GbkToUtf8(s []byte) ([]byte, error) {